+++ Changelog.md
@@ -3,3 +3,6 @@
 ### 2023/07/16
 - Added UML diagram
 - Added menu system concept written in python for fast prototyping
+
+### 2023/07/18
+- Added some logic to the menu system concept